Added visual representation of item removal
Added Remove buttons and connected them to the transaction.js file on top of Josh's existing html code. Added event listeners for the remove buttons in the transaction.js file so that they visually get removed as an element. Further along this will be the style how the user ends up calling removing an actual cart item. The transaction.js changes were made on top of Johnny's original file and format.